Stateful_Chatbot/chatbot-authentication/pyproject.toml
[project]
name = "chatbot-authentication"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.400",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.1.0",
]
